few additions

+++ src/main/java/at/fh/ima/swengs/recipebox/model/RecipeRepository.java
@@ -14,7 +14,7 @@ public interface RecipeRepository extends PagingAndSortingRepository<Recipe, Lon
     public List<Recipe> findByName(@Param("name") String name);
     public List<Recipe> findByCategory(@Param("category") String category);
     public List<Recipe> findByIngredient(@Param("ingredient") String ingredient);
-    public Recipe findById(@Param("id") int id);
+    public Recipe findById(@Param("id") long id);
     //public List<Recipe> findByNameAndIngredient(@Param("name") String name, @Param("ingredient") String ingredient);
